+++ src/transport/Bus.java
@@ -14,9 +14,14 @@ public class Bus extends Transport<DriverD> {
         return type;
     }
 
+    public CapacityBus getCapacityBus() {
+        return capacityBus;
+    }
+
     public void setType(String type) {
         this.type = type;
     }
+
 
     @Override
     public void begin() {
@@ -57,64 +62,18 @@ public class Bus extends Transport<DriverD> {
 
     }
 
-    public CapacityBus getCapacityBus() {
-        return capacityBus;
-    }
-
-    public enum CapacityBus {
-        especiallySmall(0, 10),
-        small(0, 25),
-        average(40, 50),
-        big(60, 80),
-        extraLarge(100, 120);
-        private int minCapacity, maxCapacity;
-
-
-        CapacityBus(int minCapacity, int maxCapacity) {
-            this.minCapacity = minCapacity;
-            this.maxCapacity = maxCapacity;
-        }
-
-        public int getMinCapacity() {
-            return minCapacity;
-        }
-
-        public int getMaxCapacity() {
-            return maxCapacity;
-        }
-
-        public void setMinCapacity(int minCapacity) {
-            this.minCapacity = minCapacity;
-        }
-
-        public void setMaxCapacity(int maxCapacity) {
-            this.maxCapacity = maxCapacity;
-        }
-    }
-
-
     @Override
     public String toString() {
         if(getType() == "" || getType() == null){
             return printType();
         }
         if (this.getCapacityBus().getMinCapacity() == 0) {
-            return "Transport{" +
-                    "brand='" + getBrand() + '\'' +
-                    ", model='" + getModel() + '\'' +
-                    ", engineVolume=" + getEngineVolume() +
-                    ", draiver=" + getDraiver() +
-                    '}' + " Car{" +
+            return transportPrint() + " Car{" +
                     "CapacityBus=" + getCapacityBus() + " : " + " Вместимость до "
                     + getCapacityBus().getMaxCapacity() + '}';
 
         } else {
-            return "Transport{" +
-                    "brand='" + getBrand() + '\'' +
-                    ", model='" + getModel() + '\'' +
-                    ", engineVolume=" + getEngineVolume() +
-                    ", draiver=" + getDraiver() +
-                    '}' + " Car{" +
+            return transportPrint() + " Car{" +
                     "CapacityBus=" + getCapacityBus() + " Вместимость " + getCapacityBus().getMinCapacity() + " до "
                     + getCapacityBus().getMinCapacity() + '}';
         }
